requirements.txt
autopep8==1.6.0
pycodestyle==2.8.0
pycryptodome==3.12.0
toml==0.10.2
